Generalized implementation of `D1_extra_vlines`
This commit allows the use of multiple `D1_extra_vlines`
options in the form of a list. This way, different parameters
can have different plot styles. The usage pattern is:
`D1_extra_vlines=[{[PARAMETER] : [VALUE], None : **kwargs}, ...]`
where [PARAMETER] is the name of the parameter, [VALUE] is a numeric type,
and **kwargs are keyword arguments for a `matplotlib.axes.Axes.vlines`
class, described here:
https://matplotlib.org/3.1.1/api/_as_gen/matplotlib.axes.Axes.vlines.html

Tests were also added for validation of the code.